+++ kite/src/test/resources/kite-test.sql
@@ -26,6 +26,7 @@ create table account_role (
 insert into account_role (account_id, role_id) values
 (1, 1),
 (2, 2),
+(2, 3),
 (3, 3),
 (4, 2),
 (5, 3);