+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## 1.0.5
 
+### Bug Fixes
+
+### Dependencies
+
 ### Features
 
  * Support one-to-one association query
+ * Support one-to-many association query
 
 ## 1.0.4